requirements.txt
fastapi==0.109.0
gdown
gunicorn==21.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
ipykernel==6.25.0
jsonlines==4.0.0
matplotlib
mlflow-skinny==2.9.2
numpy>=1.21.0
opencv-python
pandas==2.1.4
Pillow
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv
python-json-logger==2.0.7
python-multipart==0.0.6
streamlit
tensorflow
torch==2.1.2
torchvision==0.16.2
uvicorn[standard]==0.25.0